Fixed floor enum and refactored file structure
