+++ src/main/java/Database/DatabaseGraphController.java
@@ -6,11 +6,11 @@ import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
-import pathfinding.Floor;
-import pathfinding.HospitalEdge;
-import pathfinding.HospitalNode;
-import pathfinding.MoveAttribute;
-import pathfinding.NodeInitializer;
+import edu.wpi.teame.map.Floor;
+import edu.wpi.teame.map.HospitalEdge;
+import edu.wpi.teame.map.HospitalNode;
+import edu.wpi.teame.map.MoveAttribute;
+import edu.wpi.teame.map.NodeInitializer;
 
 public class DatabaseGraphController {
   DatabaseController DBC;
