+++ src/main/java/edu/wpi/teame/map/HospitalEdge.java
@@ -1,4 +1,4 @@
-package pathfinding;
+package edu.wpi.teame.map;
 
 public class HospitalEdge {
   String nodeOneID;
