+++ src/main/java/edu/wpi/teame/map/LocationName.java
@@ -1,4 +1,4 @@
-package pathfinding;
+package edu.wpi.teame.map;
 
 import static java.util.Objects.hash;
