+++ src/main/java/edu/wpi/teame/map/MoveAttribute.java
@@ -1,4 +1,4 @@
-package pathfinding;
+package edu.wpi.teame.map;
 
 public class MoveAttribute {
   public String nodeID;
