+++ src/main/java/edu/wpi/teame/map/NodeInitializer.java
@@ -1,4 +1,4 @@
-package pathfinding;
+package edu.wpi.teame.map;
 
 public class NodeInitializer {
