+++ src/main/java/edu/wpi/teame/map/pathfinding/AbstractPathfinder.java
@@ -1,4 +1,6 @@
-package pathfinding;
+package edu.wpi.teame.map.pathfinding;
+
+import edu.wpi.teame.map.HospitalNode;
 
 import java.util.List;
 
@@ -22,5 +24,5 @@ public abstract class AbstractPathfinder {
     return findPath(from, to);
   }
 
-  abstract List<HospitalNode> findPath(HospitalNode from, HospitalNode to);
+  public abstract List<HospitalNode> findPath(HospitalNode from, HospitalNode to);
 }
