+++ src/main/java/edu/wpi/teame/map/pathfinding/BFSPathfinder.java
@@ -1,4 +1,6 @@
-package pathfinding;
+package edu.wpi.teame.map.pathfinding;
+
+import edu.wpi.teame.map.HospitalNode;
 
 import java.util.HashMap;
 import java.util.HashSet;
